# docs/openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener API
  description: A production-ready URL shortener service
  version: 1.0.0
  contact:
    name: Duilio Fanton
    email: duilio@example.com
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - ApiKeyAuth: []
paths:
...
  /{shortCode}:
    get:
      summary: Redirect to original URL
      description: Redirect to the original URL using short code
      operationId: redirect
      parameters:
        - name: shortCode
          in: path
          required: true
          description: Short code
          schema:
            type: string
      responses:
        '302':
          description: Redirect to original URL
          headers:
            Location:
              schema:
                type: string
                example: https://google.com
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
  responses:
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'